댓글, 대댓글 좋아요/싫어요 개발

+++ common/src/main/java/com/common/entity/ReactionEntity.java
@@ -13,11 +13,11 @@ import java.time.LocalDateTime;
 public class ReactionEntity extends BaseEntity{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long reactionId;
+    private Integer reactionId;
 
     private String reaction;
 
-    public ReactionEntity(Long reactionId) {
+    public ReactionEntity(Integer reactionId) {
         this.reactionId = reactionId;
     }
